fix(frontend): extract EntityProvenance component, fix duplicate key crash
Activity tabs crashed with each_key_duplicate when a source (e.g. IPDB)
had multiple claims for the same field. Extracted the duplicated
provenance rendering into a shared EntityProvenance component and
switched to index-based each keys.
